Made changes to the mocks folder
