--- src/__tests__/taskUpdate.test.js
@@ -1,9 +1,0 @@
-import Tasks from "../modules/taskUpdate";
-
-const tasks = new Tasks;
-
-describe("Task Update actions testing", () => {
-  test("Should check for adding task list item", () => {
-    
-  });
-});
